Plugin structure unified (use sweep or timeIV plugin as example) Sweep controls moved to the sweep plugin from Keithley Section save_file temporarly moved to the plugins itself (see TODO list) Some bug fixing for timeIV Some modifications to the logging and error messaging (see README file section)
