+++ pyIVLS.ini
@@ -31,19 +31,10 @@ load = False
 name = sweep
 type = script
 function = ivsweep
-class = step
+class = sequence
 load = True
 dependencies = smu
 address = sweep
-
-[runSweep_plugin]
-name = runSweep
-type = script
-function = sequence
-class = none
-load = True
-dependencies = sweep,smu
-address = runSweep
 
 [peltierController_plugin]
 name = peltierController
@@ -78,24 +69,7 @@ load = False
 dependencies = micromanipulator,smu
 
 [sweep_settings]
-plotupdate = 1
-prescaler = 0.95
-
-[runSweep_settings]
-address = /home/ivls/
-filename = testSweep
-comment = "Some very long comment"
-samplename = test sample
-plotupdate = 1
-
-[VenusUSB2_settings]
-exposure = 156
-source = /dev/video0
-
-[Keithley2612B_settings]
 singlechannel = True
-sourcehighc = False
-drainhighc = False
 drainfollowsource = False
 channel = smuB
 inject = Voltage
@@ -125,6 +99,21 @@ drainpoints = 1
 drainlimit = 0.1
 drainnplc = 20
 draindelay = 4
+plotupdate = 1
+prescaler = 0.95
+address = /home/ivarad/Desktop
+filename = testSweep
+comment = "Some very long comment"
+samplename = test sample
+
+[VenusUSB2_settings]
+exposure = 156
+source = /dev/video0
+
+[Keithley2612B_settings]
+address = USB::0x05e6::0x2612::INSTR
+sourcehighc = False
+drainhighc = False
 
 [peltierController_settings]
 source = /dev/ttyUSB0
